Define role_name in metadata
ansible-lint now enforces that Ansible role names conform to the
requirements for Ansible collection role names, defined here:
https://docs.ansible.com/ansible/devel/dev_guide/developing_collections.html#roles-directory

Note that this excludes the use of the "-" character, which means that
any Ansible roles using the old naming scheme of ansible-role-* now
trigger an ansible-lint error.  This has caused a lot of woe:
https://github.com/ansible/ansible-lint/issues/966

Fortunately, there is an issue and a corresponding pull request with a
fix:
https://github.com/ansible/ansible-lint/issues/979
https://github.com/ansible/ansible-lint/pull/980

In short, by default ansible-lint assumes that the role name is the
name of the directory containing the role, but one is allowed to
override this behavior by setting the "role_name" entry in the role
metadata.  That will satisfy ansible-lint, and that is what I have
done here.

Also alphabetize the metadata entries by key.