Cease setting the default forward policy to accept
For hardened systems the user may well want to default policy to be
deny and create a rule to allow specific forwarded/routed traffic
through.

Also update the README to let the user know that he or she will need
to configure the forwarded/routed traffic via ufw outside of this
role.