Temporarily remove support for Fedora and Ubuntu Focal
This is due to a deficiency in cisagov/ansible-role-ufw.

For reasons I haven't been able to discern, the Fedora and Ubuntu
Focal molecule tests work fine locally but fail in GitHub Actions due
to being unable to perform some iptables operations.  I would think
this was due to some kernel module not being loaded in the underlying
host instance, except that other platforms pass just fine in GitHub
Actions.

I've already spent too much time on this, and the role works just fine
when building a Fedora AMI and we do not actually require Ubuntu Focal
support, so I will make the executive decision to remove Fedora and
Ubuntu Focal support for now.  I created issue
cisagov/ansible-role-ufw#3 to document this.