Add OCSP check to verify-cn.py
This was the first iteration of the modifications needed to have the
verify-cn.py perform OCSP checks for an OpenVPN user. It should be noted
that this code isn't currently functional. The purpose is so someone else
might be able to pick up where I left off to continue to develop this
feature.

I was able to get a mocked version of the verify-cn.py script running
locally that performed the OCSP check. When it was loaded onto the staging
OpenVPN servers it was discovered that one of the dependencies to perform
the check, the Cryptography library, on the servers was 3.3.2. The success
that I had running locally was using 41.0.7. This was an issue because
the OCSP check is trying to process multiple SINGLERESP's, 20 total, and
the outdated version of the Cryptography library only supports doing 1 at
a time. This caused a block on the development until we resolved upgrading
the OpenVPN servers from an AMI based on Debian Bullseye to Debian
Bookworm which would have the upgraded Cryptography library.