Specify a type for the value of the unit file attribute when testing
Ansible roles use YML (which has a notion of types) and systemd unit
files use INI format (which does not and treats all values as
strings).  @mcdonnnj noted in
https://github.com/cisagov/ansible-role-openvpn/pull/81#discussion_r1653429157
that there was some confusion when I wrote a (Boolean) value of
"false" in tasks/main.yml but read back a (string) value of "False" in
the test code.  To make things more explicit, I now specify a type for
the systemd unit file attributes that I am testing.
configparser.ConfigParser supports interpretation of values as
strings, integers, floating-point numbers, or Booleans.

The reason for the discrepancy that @mcdonnnj observed is that:
1. Ansible's community.general.ini_file writes the YML value "false"
to the unit file as the string "False".
2. Python's configparser.ConfigParser.get() method interprets all
values as strings, and therefore returns the string "False".

Now what happens is:
1. Ansible's community.general.ini_file writes the YML value "false"
to the unit file as the string "False".
2. Python's configparser.ConfigParser.getboolean() method interprets all
values as Booleans, and therefore interprets the string "False" as a
false Boolean value.  See
https://docs.python.org/3/library/configparser.html#configparser.ConfigParser.getboolean
for more details.